+++ docker-images/server/src/main/docker/start.sh
@@ -3,12 +3,12 @@
 APP_DIR="/opt/concord/server"
 
 MAIN_CLASS="com.walmartlabs.concord.server.Main"
-if [ "${CONCORD_COMMAND}" = "migrateDb" ]; then
+if [[ "${CONCORD_COMMAND}" = "migrateDb" ]]; then
     MAIN_CLASS="com.walmartlabs.concord.server.MigrateDB"
 fi
 
 CFG_FILE="";
-if [ ! -z "${CONCORD_CFG_FILE}" ]; then
+if [[ ! -z "${CONCORD_CFG_FILE}" ]]; then
     CFG_FILE="-Dollie.conf=${CONCORD_CFG_FILE}"
 fi
 
@@ -16,10 +16,14 @@ GC_LOG_DIR=${GC_LOG_DIR:-"${APP_DIR}/logs/gc"}
 mkdir -p ${GC_LOG_DIR}
 echo "GC logs: ${GC_LOG_DIR}"
 
-if [ -z "${CONCORD_JAVA_OPTS}" ]; then
+if [[ -z "${CONCORD_JAVA_OPTS}" ]]; then
     CONCORD_JAVA_OPTS="-Xms2g -Xmx2g -server"
 fi
 echo "CONCORD_JAVA_OPTS: ${CONCORD_JAVA_OPTS}"
+
+if [[ ! -z "${CONCORD_TMP_DIR}" ]]; then
+    CONCORD_TMP_DIR="/tmp"
+fi
 
 exec java \
 ${CONCORD_JAVA_OPTS} \